+++ policyengine_us_data/datasets/sipp/__init__.py
@@ -1,1 +1,6 @@
-from .sipp import train_tip_model, get_tip_model
+from .sipp import (
+    train_tip_model,
+    get_tip_model,
+    train_asset_model,
+    get_asset_model,
+)